% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joblog-package.R
\docType{package}
\name{joblog-package}
\alias{joblog}
\alias{joblog-package}
\title{joblog: Log cronjobs with lgr}
\description{
Log cronjobs with lgr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://s-fleck.github.io/joblog}
  \item Report bugs at \url{https://github.com/s-fleck/joblog/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
