% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joblog.R
\name{as_joblog}
\alias{as_joblog}
\alias{print.joblog}
\alias{summary.joblog}
\alias{print.joblog_summary}
\title{Joblog Objects}
\usage{
as_joblog(x)

\method{print}{joblog}(x, ...)

\method{summary}{joblog}(object, ...)

\method{print}{joblog_summary}(x, ...)
}
\arguments{
\item{x, object}{any \R object}

\item{...}{ignored}
}
\value{
\code{summary.joblog()} returns a \code{joblog_summary} \code{data.table}

\code{summary.joblog()} returns a \code{joblog_summary} \code{data.table}
}
\description{
\code{joblogs} are \code{data.tables} with custom \code{\link[=print]{print()}} and \code{\link[=summary]{summary()}} methods.
\code{as.joblog()} just adds a \code{joblog} attribute to an appropriately formatted
\code{data.table}. If you want to extract jobs from a normal log(file/\code{data.frame})
use \code{\link[=scrape_joblog]{scrape_joblog()}} instead.
}
\examples{
lg <- lgr::get_logger("test")$
  set_propagate(FALSE)$
  add_appender(lgr::AppenderBuffer$new(), "buffer")

lg$list_log(job_start("update-database"))
lg$list_log(job_finished())
lg$list_log(job_start("send-report"))
lg$list_log(job_failed("something went wrong this time"))
res <- scrape_joblog(lg$appenders$buffer$data)

print(res)
summary(res)

lg$config(NULL)  # reset logger
}
