% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_jobs.R
\name{scrape_joblog}
\alias{scrape_joblog}
\title{Scrape joblog jobs from a Json Log}
\usage{
scrape_joblog(x)
}
\arguments{
\item{file}{a `.jsonl` file as created by [lgr::AppenderJson]}
}
\value{
a `data.table`
}
\description{
Scrape joblog jobs from a Json Log
}
